fix: regeneration starting too soon
The regeneration logic for blocks was broken such that the regeneration would start already while the block was still actively damaged.
These changes will put a delay via the DelayManager for activating the regeneration.

This introduces a breaking change to ActivateRegenEvent which now requires the regen value to be specified also for the base regen.