test: adjust tests for damage and restoration scenarios to changed logic
As it is no longer possible to create these events with a negative base value we now assert that attempting to do this throws an IllegalArgumentException.

In addition, the behavior is different if the collection phase yields a negative change value in total. In that case, the triggered event is inverted, e.g., if a restoration yields a negative restoration amount the health system sends out an OnDamagedEvent.