chore!: purge delayed actions for regen cool down from DamageAuthoritySystem
Delaying (or preventing) regen should be modelled by consuming the BeforeRegenEvent in these cases. The base regen component memorizes the latest point in time when damage was inflicted on the entity to allow the BaseRegenAuthoritySystem to react accordingly.

BREAKING CHANGE: This technically breaks the (internal) API on how delayed regeneration works. It should not be visible from the outside, though.