refactor!: replace implementation in RegenAuthoritySystem
This completely swaps out the implementation of the health authority system to be more lightweight and based on trigger, collector, and notification event pattern.

The new system no longer does any complex time-keeping, neither on the level of regeneration actions as tracked in the RegenComponent, nor in determining which entities are due for a regen tick.