refactor: play damage sounds (#91)
Remove special handling of damage sounds for blocks, and instead make use of the default damage sound handling of the `DamageAuthoritySystem`.

The logic around playing damage sounds defined in `DamageSoundComponent` and/or `CharacterSoundComponent` was intertwined with (unnecessary?) dependencies between them. In addition, the `BlockDamageAuthoritySystem` had its own way of determining the damage sound of blocks and sending the trigger event to play it.